+++ pages/02_project_coach.py
@@ -358,7 +358,7 @@ with st.expander("Export Chat to PDF"):
     pdf_output_path = chatpdf.create_pdf(chat_messages=st.session_state.messages, models = models, token_counts = st.session_state.token_counts, user_name=user_name, user_course=user_course)
 
     with open(pdf_output_path, "rb") as file:
-      st.download_button(label="Download PDF", data=file, file_name=f"{user_course}_{user_name}_chatisa.pdf", mime="application/pdf", use_container_width=True)
+      st.download_button(label="Download PDF", data=file, file_name=f"{user_course}_{user_name}_chatisa.pdf", mime="application/pdf", width="stretch")
 
 
 # -----------------------------------------------------------------------------